Load translations using Webpack
The Hugo watcher recently (https://github.com/gohugoio/hugo/pull/5312) started not responding to changes to files that match the `ignoreFiles` option.
While that of course makes sense, it presents us with an issue: Previously, we needed to ignore our preact-i18n translation files since they would have otherwise cause Hugo to fail. Due to the update, changes to the translations didn't cause Hugo to rebuild anymore, making development very annoying. As we didn't find a good way to circumvent this problem, we instead decided on this change.