Closes #135: Fix company category translation
Funnily enough, this was broken both in the JS and the Hugo code, for unrelated reasons.

As for the JS code (only appears in the search results, afaik), the categories were not getting translated anymore because I missed the necessary `.map` call when moving from Algolia to Typesense. Easy fix, thankfully.
The problem in the Hugo code (company list and single pages) is a little more annoying, unfortunately. As I described in #135, these translations were working but broke when we updated from Hugo 0.47 to 0.53, with no changes to the corresponding code from our side. In short, the `(apply .Params.categories "T" ".")` just stopped working. I have not been able to find out why and have confirmed that the same issue also occurs in various different configurations of the same intent (applying the `T` (or `i18n`) function over an array of strings). Might be a Hugo bug or me missing something; I don't know. Anyway, I did find a workaround of simply iterating over the array and adding the 'manually' translated entries to a `.Scratch`. Somewhat ugly and unfortunate, but since I see no point in diagnosing this further, this will have to do.