Support more countries
Our initial decision was to only 'support' (i.e. include in the i18n widget) countries that we have suggested companies for. This commit relaxes that restriction a fair bit.

We now support any country with more than 10 (although I don't have a strong opinion on that number) records (*explicitly* including 'all' as a country).

This have two main advantages (besides the obvious one of having a longer list of supported countries, of course):

1. We have recently had quite a few submissions by users for companies from 'non-supported' countries. The experience for these users was probably quite poor. Although I have accepted those submissions, they only show up in the company list but cannot be used in the generator (or found through the search, for that matter). It is does not unlikely (or unreasonable) for a user to believe that their submission was ignored.
With this commit, a country becomes supported after just a few records, thus mostly eliminating that problem.

2. I was not at all a fan of the previous behavior of the search bars regarding country filtering. While useful, I believe it was not obvious to users that there are country filters in place and that there are companies they simply couldn't find through the search.
Thanks to the new 'all' country we now have a clear, uncluttered (no additional checkboxes etc.) and imho reasonably obvious way of disabling country filtering in the search.