Drastically improve Hugo builds times
The main culprit was that `scripts.html` had to be built for *every
single* page (all 2000 or so of them) and *every single* of those
builds iterated over all company records, counting them. Now, Hugo
certainly is fast but that still took it a couple of seconds. :D

Luckily, there is a `partialCached` function that instead just renders
the partial once (actually a couple of times, for each language etc.
but still a lot fewer times).

While we're at it, I have also disable the 'related companies' thing
in dev mode. That wasn't nearly as bad but it still saves a good
second or so and it really shouldn't be needed anyway while
developing.