Only style `.button`, not `button`
This is part of a larger rewrite of how we style elements. In the
future, we only want to style elements that have a specific class
set and not the 'raw' element.

Thus, from now on, creating a `<button class="button-primary">` is
not sufficient anymore, instead you need to make it a
`<button class="button button-primary">`. In addition,
`button-secondary` should also always be explicitly specified even
when it wouldn't be necessary.

This change is inspired by the 'eslint-plugin-jsx-a11y' rule
'anchor-is-valid'. We want to be able to differentiate between
buttons that are only functionally a button and those that should
also look like want.