Make sure translated components are wrapped in an <IntlProvider>
A component that uses preact-i18n's <Text> or <MarkupText> should
make sure that those are wrapped in an <IntlProvider> to ensure
the correct scope is used.
Otherwise, if another component includes this one, the wrong or no
translation might be loaded.

Previously, our approach to this was to have the including
component add the <IntlProvider> but this doesn't make much sense,
especially now that we are focussing more on reusing components in
different contexts.