Introduce RequestGeneratorBuilder
The big one. For the act pages feature I am working on, we need to
be able to have (customized) generators on other pages. Even with
the previously introduced components, this would still have been a
whole lot of duplicated code.

To address this, I introduced the RequestGeneratorBuilder. This new
component handles all the previous generator logic. Through it, we
can easily create custom generators using the placeholder components
without any duplicated code.