Fix signature bug for new users
While the `Request` object has a sensible default value for the
signature, that is immediately overwritten by the IdData
initialization, thus leaving us with a signature of `null` for a
new user who has neither entered a name nor painted a signature.

Several places in the code however rely on having access to
`signature.name` etc.

To solve that problem, we now return an empty signature instead of
`null` in `IdData::getSignature()`.