Try to mitigate some concurrency issues in the generator
Our generator code is really written with a synchronous workflow
in mind, yet significant parts of it are executed asynchronously.

This really starts being an issue now that we are using the
generator logic in different ways and places.

This commit seems to solve some of these issues of required_elements
set by a company being (wrongly) overwritten by an
initializeFields() call that executes later than we expect it to.