Add (very crude) dark mode for debugging only
One other immediate advantage of the new color palette is that we
can now trivially implement a 'dark mode'.

For now, this dark mode is for debugging purposes only. It is quite
helpful for spotting elements that either don't have colors assigned
or that are still using hard-coded colors.

But I am genuinely impressed with how good it looks given how
trivial the 'implementation' (literally just flipping the palette)
is. With a few tweaks, this could be pretty nice. We might want to
consider implementing a proper dark mode in the future.