Fixes #155: Correctly determine whether send/download button is enabled
We were previously checking for the existence of the 'blob_url' but
that doesn't apply for emails.

For PDFs, we now use 'download_active' which is used for the
'e.preventDefault()' in that case anyway.

For emails, we check whether we have a letter which — again is also
used for the 'preventDefault()'.