Re #209: Fix bank transfer QR codes not appearing in print window
The problem was that we were copying the transfer details using
`innerHTML` which obviously doesn't preserve the canvas data.
Now, we are also repainting the QR codes in the print window.

The print function was already specific to this use-case but it is
now even more, so it doesn't make sense to keep it in `browser.js`,
this commit moves it into the `DonationWidget.js`.

Finally, keep the QR code background pure white in the print window
as opposed to the slight off-white on the main site.