Fixes #290: Allow limited formatting in comments (#297)
* Introduce some => and .map in the comments widget

* Fixes #290: Allow limited formatting in comments

This adds support for the following, *very* basic formatting tags in
comments, similar to what the generator allows:
- <bold>This text is bold.</bold>
- <italic>This text is italic.</italic>
- <link url="https://abc.org">This text links to abc.org.</link>

The parsing is very naive, using regex, again like in the generator
code. Worryingly enough, nested tags are actually supported unlike in
the generator. I sort of regret implementing this, but it was just
toooo easy. :(

* Require urls to be https

* Update src/Components/CommentsWidget.js

Co-authored-by: Benjamin Altpeter <benjamin@oldpeter.de>

Co-authored-by: Lorenz Sieben <lorenz@datenanfragen.de>