Embed title images into act/blog entries
Some entries are listed on the act and blog site with a thumbnail of
the featured image. Some of them also reference that banner image in
the metadata section of the entry, but the image itself is not shown
on the entry site.
All entries that reference the banner ("notice" section) in their
metadata should have the correct banner image set.