Error handler: Display modal for more errors
Previously, we were only displaying the modal for errors from
our own error classes. This doesn't make much sense. There
really shouldn't be any errors on the site and if one occurs,
we would like to be notified.

We are still only displaying the modal for errors with a min.
criticality if our classes are used.