Fixes #352: Properly translate categories in the search
This commit doesn't solve the issue per se but rather removes the
mechanism that caused it as there was really no reason for that.

Let me explain: Previously, the search also queried the categories
of the records. Therefore, "schokol" matched the category "school"
(considering an allowed Levenshtein distance of 4).

For the matches, Typesense helpfully returns a "highlights"
property where our category now looks like this: "<mark>school
</mark>".

The problem was that we were also feeding these "highlighted"
results into our translate function which obviously didn't return
anything as we don't have a translation for "<mark>school</mark>".

This left me with two options: Either implement some annoying way
to translate while still preserving the highlights or just drop
searching in categories.*

The choice was quite obvious given that I don't really see any use
for searching in the categories, **especially** the untranslated
ones.

---

* Technically, just removing /<\/?mark>/ for the translation and
  then re-appending it around the whole translated word would
  have _probably_ been a good enough approximation and therefore
  another option.

  However, given that it really doesn't make sense to be able to
  search the categories anyway, this doesn't matter.