Improve sorting on company database page.
This change will make it so that accented characters in the company name are treated as equivalent Latin Alphabet characters in the sorted list on the company database page.

The change also filters out certain special characters, like various quotes, and problem characters, like Zero Width Space, so that they do not affect sorting.