Drastically improve Hugo build speed by caching related companies
Running 'hugo --templateMetrics' reveals that our build currently
(unsurprisingly) spends a disproportionate amount of time calculating
the related companies for the company single pages.
When we introduced the current algorithm in #320, we decided that the
longer build time of our custom algorithm vs. Hugo's built-in one is
worth it, given the vastly greater improvement in deploy times it
offers.

However, since then we have not only added plenty of new companies
but also another language, which had a significant impact on the
build times. Whereas the build previously took around 18s on my
(rather fast) machine, it now takes 45s. This effect is of course
greatly amplified in CI, where we are running on much slower
machines.
On Netlify, we have seen the Hugo build take 12 minutes and more.
Thus, there is obviously a great need to improve on this.

Luckily, there is one more trick up our sleeves before we need to
think about a more clever algorithm. We can move the calculation into
a partial and make use of Hugo's 'partialCached' function which
caches the result of the partial so it doesn't have to be recomputed,
as the name implies. This function also accepts optional VARIANT
arguments that we need to make sure that not all companies have the
same related companies. We can simply pass both the categories and
relevant countries as variants as they are the only parameters that
influence the algorithm.
As such, the number of necessary calculations is *greatly* reduced.
Whereas the 'company/single.html' template is currently executed 9996
times (once for each language and company), our new
'partials/related-companies.html' partial only needs to be executed
1647 times. Still a lot but a clear improvement.

In effect, this brings the build time down to 12s on my machine while
maintaining the exact same results.