Closes #502: Lazy load images below the fold
We are cheating a little here. We have no way of knowing which images
are above or below the fold from Hugo. Thus, we make a (reasonable)
assumption: Featured images are always above the fold, while all other
images are below the fold.

For that, we introduce the `featuredImg` shortcode which defaults to
eager loading, while the `img` shortcode now defaults to lazy loading.
This new shortcode has the added benefit of avoiding having to
duplicate the feature image name in the Markdown, which was actually a
little annoying.

If necessary, it is still possible to override this behaviour for other
images using the `loading` parameter of the `img` shortcode.