Inline fonts CSS from CDN
Google was (rightfully) complaining that the CSS files from
static.dacdn.de for loading the fonts were adding too much overhead.

This is mostly true for the first of the two (fira-go-1001.css) since
a connection to an entirely new origin had to be established first.

Duplicating this into website shouldn't be too much of an issue since
this CSS rarely ever changes.