Don't delete signature if the SignatureInput JS executes too quickly
This bug was caused by the switch to Preact X, where `setState()`
isn't synchronous anymore.

The `SignatureInput`'s state starts with `isEmpty === true` and runs
the `drawSignature()` function initially to, well, draw the stored
signature (if any) to the canvas. This function then sets `isEmpty`
to `false` and calls `handleChange()` which will store the signature
or erase it if `isEmpty === true`.

And that's exactly where the problem lies. Previously, the `setState()`
call would block and only call `handleChange()` after the state had
been updated. But that isn't the case anymore with Preact X and as
such, the signature was cleared if `handleChange()` was called too
quickly and the state hadn't been updated yet.

The solution is to simply move the `handleChange()` call to the a
callback in the optional second argument of `setState()` which is only
executed after the state has been updated.