Closes #504: Fix cut-off 'required' asterisk
The view box of the SVG was not big enough.

Also change font from FiraGO to Arial as, at least for me, FiraGO
isn't available in the SVG and instead replaced by a default font
that looks quite ugly here.
Arial should be available on just about every system and also ensures
that the asterisk has a consistent height and isn't cut off (again)
depending on the default font per system.