Fix featured image in English 'direct markting objection' post (#690)
The English site was displaying the Spanish image. That was because
we match the featured image by `*<featured_image>*`. The one for the
English post was called `sample-letter-gdpr-direct-marketing-objection`,
and the Spanish one `sample-letter-gdpr-direct-marketing-objection-es`.

So, when we specified `sample-letter-gdpr-direct-marketing-objection`
for `featured_image` in the English post, it matched both and just so
happened to match the Spanish one first.