Related companies: Reduce sorting effort
For a company, the template previously had to sort all potential
related companies by their score. This list can get quite long since
it includes every company with at least one category in common.

However, in the end, we only pick the top 5. So, it doesn't make sense
to add companies with a score lower than the current lowest score to
the list if we already have five or more companies in the list.

While this does improve the speed a little, it doesn't do much. It does
however bring us close to the speed we would get without any sorting.

On my machine, previously a build took 27691 ms with an average
duration for this template of 129.618492ms.
With this patch, this comes down to 27077 ms and 291.796755ms.