Related companies: Only compute common countries if necessary
A company is only considered as a candidate related company if it has
a least one category in common with the company in question. Thus, it
doesn't make sense to do any other calculations (like the number of
countries in common) otherwise.

With this patch, the build time on my machine drops from 27095 ms to
23022 ms. Quite a nice improvement for moving one line down a tad. :)