Related companies: Exclude disqualified companies early
We know right from the start that we will never add ourselves or
nsfw companies to the list of related company candidates, so it doesn't
make sense to do any calculations for this.

Obviously, this doesn't help much, bringing the build time from
23022 ms to 22923 ms on my machine.
This might change should we start adding significantly more nsfw
companies. :D