Optimize deploy script by only renaming once
With more languages and companies being added, our deploy script has
been getting slower and slower for a while. That makes sense since we
have to copy and rename each record for each language. With currently
more than 3,000 companies and 8 languages, that is quite a lot of work.

Today, I happened to notice that the copying of the files is super fast
while the renaming takes a long time. Taking a closer look at the
script, that shouldn't be too surprising: We're spawning a separate
shell for each rename operation, ~24,000 shells in total.

There's a trivial optimization to be had here: Only do the renaming
once and _then_ copy the already renamed files. While still not great
(since we still spawn ~3,000 shells), that is a huge improvement.

On my machine, the previous version of the deploy script took 56s
excluding the "Running Webpack and Hugo…" steps. With this change,
it only takes 12s.

And since we're literally just rearranging code, this should also be
fine in CI, unlike our earlier tries with GNU parallel.