Suggest: Hide required elements and comments (#1103)
We never know whether people are just guessing for the
required elements and usually end up removing them anyway
for that reason.

And as for the comments: Records should only have comments
in very rare cases. This field just encouraged people to
leave the kinds of comments that should be left on the
company pages instead.