Fixes #1139: Reverse default sort order in RequestList (#1221)
I deliberately didn't change the sort order for the exports (CSV and
ICS). I think for those, old to new makes more sense.

I've added a prop to RequestList to determine the sort order. In the
future, we might want to add a UI toggle for that. But this order seems
much more sensible to me.